refactor: Restructure imports and remove unused router.py file; update related modules
