+++ .gitignore
@@ -170,3 +170,4 @@ data/test_connections_*.json
 *.txt
 *.jsonl
 nul
+*.gemini
