+++ src/ai/router/__init__.py
@@ -1,7 +1,7 @@
 """
 Router module - staged conversation router for ICC agent.
 """
-from src.ai.router.router import handle_turn
+from ai.router.router_orchestrator import handle_turn
 from src.ai.router.memory import Memory, create_memory
 from src.ai.router.context.stage_context import Stage
